refactor: extract BudgetServiceInterface, lock BudgetService final readonly, isolate DB aggregation
Slice 11 — completes the audit's "Final classes" row at 10/10 and
addresses the last service the audit explicitly listed as needing
locking.

### Architectural change

`BudgetServiceTest` previously mocked `aggregateWindowUsage()` via
anonymous classes that extended `BudgetService` — that pattern is
incompatible with `final`. Rather than making the production class
non-extension-friendly, the SQL aggregation moves out of the service
and into its own collaborator:

- New `Classes/Service/Budget/BudgetUsageWindowsInterface.php` —
  contract for "give me daily + monthly usage rollups".
- New `Classes/Service/Budget/UserBudgetUsageWindows.php` —
  `final readonly` SQL implementation; the conditional-SUM single-
  roundtrip query that lived in `BudgetService::aggregateWindowUsage()`
  moved here verbatim.
- `BudgetService::__construct()` now takes
  `(UserBudgetRepository, BudgetUsageWindowsInterface)` instead of
  `(UserBudgetRepository, ConnectionPool)`. The service itself is
  shorter and focused — budget evaluation only.

### Lock + interface

- New `Classes/Service/BudgetServiceInterface.php` exposing the single
  public method `check()`.
- `BudgetService` is now `final readonly` and implements the interface.
- `BudgetMiddleware` constructor typehint updated to the interface.
- `Configuration/Services.yaml` aliases:
    - `BudgetServiceInterface → BudgetService`
    - `BudgetUsageWindowsInterface → UserBudgetUsageWindows`

### Test refactor

- `BudgetServiceTest::makeService()` now stubs
  `BudgetUsageWindowsInterface` instead of subclassing the service. No
  more anonymous subclasses.
- The "single aggregation call" test uses a `MockObject` that captures
  the daily/monthly timestamp arguments via `willReturnCallback()`,
  preserving the original assertion that exactly one DB roundtrip is
  made and the monthly window starts before the daily one.
- `BudgetMiddlewareTest` mocks the new interface instead of the
  concrete class.

### Verified

PHP 8.4: cgl clean, PHPStan level 10 clean, Rector dry-run clean,
3219 unit tests green.

### CHANGELOG

Added a `BudgetService` entry under `[Unreleased] § BREAKING` noting
the constructor-signature change for direct instantiators (autowired
consumers are unaffected via the new aliases).
